confirmation delete page create

+++ ration/views.py
@@ -174,12 +174,14 @@ def recipe_update(request, recipe_id):
   context = {'recipe': recipe, 'form': form, 'ingredients': ingredients, 'ingredient_list': ingredient_list, 'SIZES':SIZES }
   return render(request, 'recipes/create.html', context)
 
+@login_required
 def delete_recipe(request, recipe_id):
   recipe=Recipe.objects.get(id=recipe_id)
-  if recipe.user_id == request.user.id:
-    recipe.delete()
-  return redirect('index')
-
+  if request.method == 'POST':
+    if recipe.user_id == request.user.id:
+      recipe.delete()
+      return redirect('index')
+  return render(request, 'recipes/confirm_delete.html', {'recipe': recipe})
 
 @login_required
 def add_review(request, recipe_id):
